requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated_types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
async_timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset_normalizer==3.4.1
click==8.1.8
coverage==7.7.1
dataclasses_json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docstring_parser==0.16
docutils==0.21.2
exceptiongroup==1.2.2
execnet==2.1.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.0
googleapis_common_protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx_sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.6.0
langchain==0.3.21
langchain_community==0.3.20
langchain_core==0.3.48
langchain_openai==0.3.10
langchain_text_splitters==0.3.7
langfuse==3.2.2
langgraph==0.3.20
langgraph_checkpoint==2.0.23
langgraph_prebuilt==0.1.6
langgraph_sdk==0.1.59
langsmith==0.3.18
llama_cloud==0.1.16
llama_cloud_services==0.6.8
llama_index==0.12.25
llama_index_agent_openai==0.4.6
llama_index_cli==0.4.1
llama_index_core==0.12.25
llama_index_embeddings_openai==0.3.1
llama_index_indices_managed_llama_cloud==0.6.9
llama_index_llms_openai==0.3.27
llama_index_multi_modal_llms_openai==0.4.3
llama_index_program_openai==0.3.1
llama_index_question_gen_openai==0.3.0
llama_index_readers_file==0.4.7
llama_index_readers_llama_parse==0.4.0
llama_parse==0.6.4.post1
markdown_it_py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
more_itertools==10.6.0
multidict==6.2.0
mypy_extensions==1.0.0
nest_asyncio==1.6.0
networkx==3.4.2
nh3==0.2.21
nltk==3.9.1
numpy==1.26.4
ollama==0.4.7
openai==1.68.2
orjson==3.10.16
ormsgpack==1.9.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
pkginfo==1.10.0
platformdirs==4.3.7
pluggy==1.5.0
portalocker==3.1.1
propcache==0.3.0
protobuf==5.29.4
psycopg2_binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydantic_settings==2.8.1
pygments==2.19.1
PyJWT==2.10.1
pypdf==5.4.0
pytest==8.3.5
pytest_asyncio==0.26.0
pytest_repeat==0.9.3
pytest_rerunfailures==12.0
pytest_xdist==3.6.1
python_dateutil==2.9.0.post0
python_dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
readme_renderer==44.0
regex==2024.11.6
requests==2.32.3
requests_toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
SecretStorage==3.3.3
sentry_sdk==2.24.1
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.39
striprtf==0.0.26
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
twine==5.1.1
typer==0.15.2
typing_extensions==4.12.2
typing_inspect==0.9.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
watchdog==6.0.0
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
uvicorn==0.34.0
fastapi==0.115.12
gunicorn==23.0.0
mcp==1.9.2
langchain_mcp_adapters==0.1.7
holidays==0.77
python-docx==1.2.0
htmldocx==0.0.6
pyxlsb==1.0.10
markdown==3.7
openpyxl==3.1.5
